+++ Common/Src/datetypes/quaternion.cpp
@@ -7,19 +7,19 @@ Quaternion::Quaternion() {
     this->w = 1;
 }
 
-Quaternion::Quaternion(double x, double y, double z, double w) {
+Quaternion::Quaternion(float x, float y, float z, float w) {
     this->x = x;
     this->y = y;
     this->z = z;
     this->w = w;
 }
 
-double Quaternion::getX() { return this->x; }
-double Quaternion::getY() { return this->y; }
-double Quaternion::getZ() { return this->z; }
-double Quaternion::getW() { return this->w; }
+float Quaternion::getX() { return this->x; }
+float Quaternion::getY() { return this->y; }
+float Quaternion::getZ() { return this->z; }
+float Quaternion::getW() { return this->w; }
 
-void Quaternion::setX(double x) { this->x = x; }
-void Quaternion::setY(double y) { this->y = y; }
-void Quaternion::setZ(double z) { this->z = z; }
-void Quaternion::setW(double w) { this->w = w; }
+void Quaternion::setX(float x) { this->x = x; }
+void Quaternion::setY(float y) { this->y = y; }
+void Quaternion::setZ(float z) { this->z = z; }
+void Quaternion::setW(float w) { this->w = w; }
